[refactor/#170] 쏠마크 no result 컴포넌트화
