+++ src/components/Solmark/SolmarkContentSollect.tsx
@@ -3,11 +3,10 @@ import React, { useEffect, useState } from 'react';
 import { fetchSolmarkSollect } from '../../api/solmarkApi';
 import SollectList from '../Sollect/SollectList';
 import { SollectPhotoType } from '../../types';
-import arrow from '../../assets/arrow.svg';
 import { useNavigate } from 'react-router-dom';
+import SolmarkNoResult from './SolmarkNoResult';
 
 const SolmarkContentSollect = () => {
-  const navigate = useNavigate();
   const { data } = useQuery({
     queryKey: ['solmarkSollects'],
     queryFn: () => fetchSolmarkSollect(),
@@ -31,17 +30,7 @@ const SolmarkContentSollect = () => {
           })}
         />
       ) : (
-        <div className='text-center py-250'>
-          <p className='font-bold text-primary-950 mb-5'>
-            아직 저장된 쏠렉트가 없어요!
-          </p>
-          <p
-            className='flex justify-center text-sm text-primary-700 underline items-center'
-            onClick={() => navigate('/sollect')}>
-            쏠렉트 보러 가기
-            <img src={arrow} alt='' className='w-24 h-24' />{' '}
-          </p>
-        </div>
+        <SolmarkNoResult type='sollect' />
       )}
     </div>
   );
